Correct some bug and add some first unit test to the continuous action (but not enough)

+++ gegelatilib/src/learn/learningEnvironment.cpp
@@ -47,7 +47,7 @@ bool Learn::LearningEnvironment::isCopyable() const
     return false;
 }
 
-void Learn::LearningEnvironment::doAction(uint64_t actionID)
+void Learn::LearningEnvironment::doAction(double actionID)
 {
     if (vectActions.size() > 1) {
         throw std::runtime_error(
